requirements.txt
streamlit 1.44.1
Pillow 11.2.1